+++ sdk/python/pulumi_alicloud/edas/deploy_group.py
@@ -5,7 +5,7 @@
 import warnings
 import pulumi
 import pulumi.runtime
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Mapping, Optional, Sequence, Union
 from .. import _utilities, _tables
 
 __all__ = ['DeployGroup']
@@ -79,7 +79,7 @@ class DeployGroup(pulumi.CustomResource):
             opts: Optional[pulumi.ResourceOptions] = None,
             app_id: Optional[pulumi.Input[str]] = None,
             group_name: Optional[pulumi.Input[str]] = None,
-            group_type: Optional[pulumi.Input[float]] = None) -> 'DeployGroup':
+            group_type: Optional[pulumi.Input[int]] = None) -> 'DeployGroup':
         """
         Get an existing DeployGroup resource's state with the given name, id, and optional extra
         properties used to qualify the lookup.
@@ -89,7 +89,7 @@ class DeployGroup(pulumi.CustomResource):
         :param pulumi.ResourceOptions opts: Options for the resource.
         :param pulumi.Input[str] app_id: The ID of the application that you want to deploy.
         :param pulumi.Input[str] group_name: The name of the instance group that you want to create.
-        :param pulumi.Input[float] group_type: The type of the instance group that you want to create. Valid values: 0: Default group. 1: Phased release is disabled for traffic management. 2: Phased release is enabled for traffic management.
+        :param pulumi.Input[int] group_type: The type of the instance group that you want to create. Valid values: 0: Default group. 1: Phased release is disabled for traffic management. 2: Phased release is enabled for traffic management.
         """
         opts = pulumi.ResourceOptions.merge(opts, pulumi.ResourceOptions(id=id))
 
@@ -118,7 +118,7 @@ class DeployGroup(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="groupType")
-    def group_type(self) -> pulumi.Output[float]:
+    def group_type(self) -> pulumi.Output[int]:
         """
         The type of the instance group that you want to create. Valid values: 0: Default group. 1: Phased release is disabled for traffic management. 2: Phased release is enabled for traffic management.
         """
